Login toegevoegd

+++ controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -40,16 +41,26 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path.replace("\\", "/");
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
-        title: title, content: content, imageUrl: imageUrl, creator: { name: 'Quinten' }
+        title: title, content: content, imageUrl: imageUrl, creator: req.userId
     });
     post.save().then(result => {
+        return User.findById(req.userId);
+    })
+    .then(user => {
+        creator = user;
+        user.posts.push(post);
+        return user.save();
+    })
+    .then(result => {
         res.status(201).json({
             message: 'Post success',
-            post: result
+            post: post,
+            creator: {_id: creator._id, name: creator.name}
         });
-
-    }).catch(err => {
+    })
+    .catch(err => {
         if (!err.statusCode) {
             err.statusCode = 500;
         }
@@ -102,6 +113,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId){
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -130,8 +146,20 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId){
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
             clearImage(post.imageUrl);
             return Post.findByIdAndDelete(postId);
+        })
+        .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            user.save();
         })
         .then(result => {
             res.status(200).json({ message: 'Deleted post' })
